add booking, active table api

+++ PoolLab.Application/Services/BilliardTableService.cs
@@ -21,14 +21,92 @@ namespace PoolLab.Application.Interface
         private readonly IUnitOfWork _unitOfWork;
         private readonly IQRCodeGenerate _qRCodeGenerate;
         private readonly IAzureBlobService _azureBlobService;
+        private readonly IBidaTypeAreaService _bidaTypeAreaService;
+        private readonly IConfigTableService _configTableService;
+        private readonly IBookingService _bookingService;
 
-        public BilliardTableService(IMapper mapper, IUnitOfWork unitOfWork, IQRCodeGenerate qRCodeGenerate, IAzureBlobService azureBlobService)
+        public BilliardTableService(IMapper mapper, IUnitOfWork unitOfWork, IQRCodeGenerate qRCodeGenerate, IAzureBlobService azureBlobService, IBidaTypeAreaService bidaTypeAreaService, IConfigTableService configTableService, IBookingService bookingService)
         {
             _mapper = mapper;
             _unitOfWork = unitOfWork;
             _qRCodeGenerate = qRCodeGenerate;
             _azureBlobService = azureBlobService;
+            _bidaTypeAreaService = bidaTypeAreaService;
+            _configTableService = configTableService;
+            _bookingService = bookingService;
         }
+
+
+        public async Task<string?> ActivateTable(ActiveTable activeTable)
+        {
+            try
+            {
+                var now = DateTime.UtcNow;
+                var Time = TimeOnly.FromDateTime(now);
+                var check = await _unitOfWork.BilliardTableRepo.CheckTableAvailable(activeTable.BilliardTableID, activeTable.CustomerID, now);
+                if (check != null)
+                {
+                    var config = await _configTableService.GetConfigTableByName();
+                    if (config == null)
+                    {
+                        return "Không tìm thấy cấu hình!";
+                    }
+                    if (check.CustomerId == activeTable.CustomerID)
+                    {
+                        if (Time >= check.TimeStart && Time <= check.TimeStart.Value.AddMinutes((double)config.TimeDelay))
+                        {
+                            var book = await _bookingService.UpdateStatusBooking(check.Id, "Hoàn Thành");
+                            if (book != null)
+                            {
+                                return book;
+                            }
+                        }
+                        else if (Time < check.TimeStart)
+                        {
+                            return "Bạn đến quá sớm, hãy quay lại khi tới giờ!";
+                        }
+                        else
+                        {
+                            var book = await _bookingService.UpdateStatusBooking(check.Id, "Đã Huỷ");
+                            if (book != null)
+                            {
+                                return book;
+                            }
+                        }
+                    }
+                    else
+                    {
+                        if (Time <= check.TimeStart && Time >= check.TimeStart.Value.AddMinutes((double)-config.TimeHold))
+                        {
+                            return "Bàn chơi này đã được đặt trước!";
+                        }
+                        else
+                        {
+                            return check.TimeStart.ToString();
+                        }
+                    }
+                }
+                var table = await _unitOfWork.BilliardTableRepo.GetByIdAsync(activeTable.BilliardTableID);
+                if (table == null)
+                {
+                    return "Không tìm thấy bàn chơi này!";
+                }
+                table.Status = "Có Khách";
+                _unitOfWork.BilliardTableRepo.Update(table);
+                var result = await _unitOfWork.SaveAsync() > 0;
+                if (!result)
+                {
+                    return "Kích hoạt thất bại!";
+                }
+
+                return null;
+            }
+            catch (DbUpdateException)
+            {
+                throw;
+            }
+        }
+
 
         public async Task<string?> AddNewTable(NewBilliardTableDTO newBilliardTableDTO)
         {
@@ -51,10 +129,19 @@ namespace PoolLab.Application.Interface
                 {
                     return "Tạo QRCode thất bại!";
                 }
-                bida.Qrcode = await _azureBlobService.UploadQRCodeImageAsync("qrcode",qrcode);
+                bida.Qrcode = await _azureBlobService.UploadQRCodeImageAsync("qrcode", qrcode);
                 if (bida.Qrcode == null)
                 {
                     return "Tệp tải lên thất bại!";
+                }
+                NewTypeAreaDTO bidaTypeDTO = new NewTypeAreaDTO();
+                bidaTypeDTO.AreaID = bida.AreaId;
+                bidaTypeDTO.BilliardTypeID = bida.BilliardTypeId;
+                bidaTypeDTO.StoreID = bida.StoreId;
+                var dup = await _bidaTypeAreaService.AddNewBidaTypeArea(bidaTypeDTO);
+                if (dup != null)
+                {
+                    return "Tạo mới thất bại!";
                 }
                 await _unitOfWork.BilliardTableRepo.AddAsync(bida);
                 var result = await _unitOfWork.SaveAsync() > 0;
@@ -75,9 +162,14 @@ namespace PoolLab.Application.Interface
             try
             {
                 var table = await _unitOfWork.BilliardTableRepo.GetByIdAsync(Id);
-                if(table == null)
+                if (table == null)
                 {
                     return "Không tìm thấy bàn chơi này!";
+                }
+                var typeArea = await _bidaTypeAreaService.DeleteBidaTypeArea(table.BilliardTypeId, table.AreaId, table.StoreId);
+                if (typeArea != null)
+                {
+                    return typeArea;
                 }
                 _unitOfWork.BilliardTableRepo.Delete(table);
                 var result = await _unitOfWork.SaveAsync() > 0;
@@ -86,7 +178,8 @@ namespace PoolLab.Application.Interface
                     return "Xoá thất bại!";
                 }
                 return null;
-            }catch(DbUpdateException)
+            }
+            catch (DbUpdateException)
             {
                 throw;
             }
@@ -114,7 +207,7 @@ namespace PoolLab.Application.Interface
                 query = query.Where(x => x.PriceId.Equals(bidaTableFilter.PriceId));
 
             if (!string.IsNullOrEmpty(bidaTableFilter.Status))
-                query = query.Where(x => x.Status.Contains(bidaTableFilter.Status, StringComparison.OrdinalIgnoreCase));           
+                query = query.Where(x => x.Status.Contains(bidaTableFilter.Status, StringComparison.OrdinalIgnoreCase));
 
             //Sorting
             if (!string.IsNullOrEmpty(bidaTableFilter.SortBy))
@@ -187,13 +280,13 @@ namespace PoolLab.Application.Interface
                 }
                 return null;
             }
-            catch(DbUpdateException) 
+            catch (DbUpdateException)
             {
-               throw;
+                throw;
             }
         }
 
-        public async Task<string?> UpdateStatusTable(Guid Id, string statusTableDTO)
+        public async Task<string?> UpdateStatusTable(Guid Id, UpdateStatusTableDTO statusTableDTO)
         {
             try
             {
@@ -202,11 +295,7 @@ namespace PoolLab.Application.Interface
                 {
                     return "Không tìm thấy bàn chơi này!";
                 }
-                if (statusTableDTO.Equals("Kích hoạt"))
-                {
-
-                }
-                table.Status = statusTableDTO;
+                table.Status = statusTableDTO.Status;
                 _unitOfWork.BilliardTableRepo.Update(table);
                 var result = await _unitOfWork.SaveAsync() > 0;
                 if (!result)
@@ -215,7 +304,7 @@ namespace PoolLab.Application.Interface
                 }
                 return null;
             }
-            catch(DbUpdateException)
+            catch (DbUpdateException)
             {
                 throw;
             }
